move ComponentsConfigurationUtils to Utils directory
